+++ demo/site/src/blocks/LinkBlock.tsx
@@ -1,5 +1,17 @@
+<<<<<<< HEAD
 "use client";
 import { ExternalLinkBlock, OneOfBlock, PropsWithData, SupportedBlocks, withPreview } from "@comet/cms-site";
+=======
+import {
+    DamFileDownloadLinkBlock,
+    ExternalLinkBlock,
+    InternalLinkBlock,
+    OneOfBlock,
+    PropsWithData,
+    SupportedBlocks,
+    withPreview,
+} from "@comet/cms-site";
+>>>>>>> main
 import { LinkBlockData } from "@src/blocks.generated";
 import { NewsLinkBlock } from "@src/news/blocks/NewsLinkBlock";
 import * as React from "react";
@@ -21,6 +33,11 @@ const supportedBlocks: SupportedBlocks = {
         <NewsLinkBlock data={props} title={title}>
             {children}
         </NewsLinkBlock>
+    ),
+    damFileDownload: ({ children, title, ...props }) => (
+        <DamFileDownloadLinkBlock data={props} title={title}>
+            {children}
+        </DamFileDownloadLinkBlock>
     ),
 };