+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @comet/admin-react-select
 
+## 6.12.0
+
+### Patch Changes
+
+-   Updated dependencies [16ffa7be9]
+    -   @comet/admin@6.12.0
+
+## 6.11.0
+
+### Patch Changes
+
+-   Updated dependencies [8e3dec523]
+    -   @comet/admin@6.11.0
+
 ## 6.10.0
 
 ### Patch Changes